adjusted imports, added commenting
