+++ frontend/src/screens/Sign-Up/SignUpHelpers.ts
@@ -1,13 +1,12 @@
-export const handleSubmit = async (signupDetails, addUserFunction, onSuccess, onFailure) => {
-    try {
-      const apiResponse = await addUserFunction(signupDetails);
-      if (apiResponse && apiResponse.status === 200) {
-        onSuccess(); 
-      } else {
-        onFailure('Internal server error, please try again later'); 
+    //Function uses regex to validate the password string meets necessary criteria
+    //Min 8 characters, 1 uppercase, 1 lowercase, 1 special char
+    export const validatePassword = (password: string) => {
+        const validations = {
+          minLength: password.length >= 8,
+          hasLowerCase: /[a-z]/.test(password),
+          hasUpperCase: /[A-Z]/.test(password),
+          hasSpecialChar: /[!@#$%^&*(),.?":{}|<>]/.test(password),
+        }
+    
+        return validations;
       }
-    } catch (err) {
-      console.error(err);
-      onFailure(err.message); 
-  };
-}
